Reorganize file fetcher into utils
Arvados-DCO-1.1-Signed-off-by: Stephen Smith <stephen@curii.com>
